+++ src/base/PhpBlock.php
@@ -67,4 +67,47 @@ abstract class PhpBlock extends InternalBaseBlock implements PhpBlockInterface, 
         $this->injectorSetup();
         return $this->admin();
     }
+    
+    /**
+     * Save the registered assets form block to the cache.
+     *
+     * @since 1.0.5
+     */
+    public function onRegister()
+    {
+        if ($this->getIsCacheEnabled()) {
+            $phpBlockView = $this->getView();
+    
+            $blockId = $this->getEnvOption('id');
+    
+            $cacheKeyAssets = ['blockassets', $blockId];
+            $cacheKeyAssetBundles = ['blockassetbundles', $blockId];
+            
+            $assets = Yii::$app->cache->getOrSet($cacheKeyAssets, [$phpBlockView, 'getBlockAssets'], $this->getCacheExpirationTime());
+            $assetBundles = Yii::$app->cache->getOrSet($cacheKeyAssetBundles, [$phpBlockView, 'getAssetBundleNames'], $this->getCacheExpirationTime());
+    
+            /**
+             * @todo i think this is not need because in PhpBlockView::init() the EVENT_AFTER_RENDER is listen also.
+             */
+            PhpBlockView::registerToAppView($assets, $assetBundles);
+        }
+    }
+    
+    /**
+     * Load the block assets from cache and register to the app view.
+     *
+     * @since 1.0.5
+     */
+    public function onRegisterFromCache()
+    {
+        $blockId = $this->getEnvOption('id');
+    
+        $cacheKeyAssets = ['blockassets', $blockId];
+        $cacheKeyAssetBundles = ['blockassetbundles', $blockId];
+    
+        $assets = Yii::$app->cache->get($cacheKeyAssets) ?: [];
+        $assetBundles = Yii::$app->cache->get($cacheKeyAssetBundles) ?: [];
+    
+        PhpBlockView::registerToAppView($assets, $assetBundles);
+    }
 }