+++ tests/data/blocks/UnitTestBlock.php
@@ -6,6 +6,14 @@ use luya\cms\base\InternalBaseBlock;
 
 abstract class UnitTestBlock extends InternalBaseBlock
 {
+    public function onRegister()
+    {
+    }
+    
+    public function onRegisterFromCache()
+    {
+    }
+    
     public function renderAdmin()
     {
     }